Deprecate product-summary-brand
