--- react/ProductSummaryBrand.js
@@ -1,3 +1,0 @@
-import ProductSummaryBrand from './components/ProductSummaryBrand/ProductSummaryBrand'
-
-export default ProductSummaryBrand
